Interchangeable loop constructs
As some loop constructs perform better in some browsers than others,
they're now interchangeable.
Developers can now optionally pass in a JSON response from
useragentstring.com to activate the most performant loop construct for
the end user's browser.
I'll develop an optional plugin to make an AJAX call for that JSON, but
it's recommended to do that on the server side to save the client an
additional http request.